+++ example/lib/pages/markers_page.dart
@@ -16,18 +16,18 @@ class MarkersPage extends StatefulWidget {
 
 class MarkersPageState extends State<MarkersPage> {
   final controller = MapController(
-    location: LatLng(35.68, 51.41),
+    location: const LatLng(35.68, 51.41),
   );
 
   final markers = [
-    LatLng(35.674, 51.41),
-    LatLng(35.678, 51.41),
-    LatLng(35.682, 51.41),
-    LatLng(35.686, 51.41),
+    const LatLng(35.674, 51.41),
+    const LatLng(35.678, 51.41),
+    const LatLng(35.682, 51.41),
+    const LatLng(35.686, 51.41),
   ];
 
   void _gotoDefault() {
-    controller.center = LatLng(35.68, 51.41);
+    controller.center = const LatLng(35.68, 51.41);
     setState(() {});
   }
 
@@ -105,7 +105,7 @@ class MarkersPageState extends State<MarkersPage> {
             (pos) => _buildMarkerWidget(pos, Colors.red),
           );
 
-          final homeLocation = transformer.toOffset(LatLng(35.68, 51.42));
+          final homeLocation = transformer.toOffset(const LatLng(35.68, 51.42));
 
           final homeMarkerWidget =
               _buildMarkerWidget(homeLocation, Colors.black, Icons.home);