+++ example/lib/pages/twilight_page.dart
@@ -20,7 +20,7 @@ class TwilightPage extends StatefulWidget {
 
 class TwilightPageState extends State<TwilightPage> {
   final controller = MapController(
-    location: LatLng(35.68, 51.41),
+    location: const LatLng(35.68, 51.41),
     zoom: 4,
   );
 
@@ -36,7 +36,7 @@ class TwilightPageState extends State<TwilightPage> {
   }
 
   void _gotoDefault() {
-    controller.center = LatLng(35.68, 51.41);
+    controller.center = const LatLng(35.68, 51.41);
     setState(() {});
   }
 
@@ -93,14 +93,14 @@ class TwilightPageState extends State<TwilightPage> {
               civil.polyline.map((e) => transformer.toOffset(e)).toList();
 
           if (civil.delta < 0) {
-            var p1 = transformer.toOffset(LatLng(90, -180));
-            var p2 = transformer.toOffset(LatLng(90, 180));
+            var p1 = transformer.toOffset(const LatLng(90, -180));
+            var p2 = transformer.toOffset(const LatLng(90, 180));
 
             polyline.insert(0, p1);
             polyline.add(p2);
           } else {
-            var p1 = transformer.toOffset(LatLng(-90, -180));
-            var p2 = transformer.toOffset(LatLng(-90, 180));
+            var p1 = transformer.toOffset(const LatLng(-90, -180));
+            var p2 = transformer.toOffset(const LatLng(-90, 180));
 
             if (p1.dy > big.height) {
               p1 = Offset(p1.dx, big.height);