+++ example/lib/pages/vector_map_page.dart
@@ -13,11 +13,11 @@ class VectorMapPage extends StatefulWidget {
 
 class VectorMapPageState extends State<VectorMapPage> {
   final controller = MapController(
-    location: LatLng(35.68, 51.41),
+    location: const LatLng(35.68, 51.41),
   );
 
   void _gotoDefault() {
-    controller.center = LatLng(35.68, 51.41);
+    controller.center = const LatLng(35.68, 51.41);
     setState(() {});
   }