+++ example/lib/utils/twilight_painter.dart
@@ -33,5 +33,6 @@ class TwilightPainter extends CustomPainter {
   }
 
   @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) => true;
+  bool shouldRepaint(covariant TwilightPainter oldDelegate) =>
+      oldDelegate.border != border;
 }